requirements.txt
pygame==2.5.2
numpy==1.26.0
Pillow==10.0.1
deap==1.4.1
scikit-learn==1.3.2